% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{as.data.frame.phate}
\alias{as.data.frame.phate}
\title{Convert a PHATE object to a data.frame}
\usage{
\method{as.data.frame}{phate}(phate, ...)
}
\arguments{
\item{phate}{A fitted PHATE object}

\item{...}{Arguments for as.data.frame()}
}
\description{
Returns the embedding matrix with column names PHATE1 and PHATE2
}
