% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{as.matrix.phate}
\alias{as.matrix.phate}
\title{Convert a PHATE object to a matrix}
\usage{
\method{as.matrix}{phate}(x, ...)
}
\arguments{
\item{x}{A fitted PHATE object}

\item{...}{Arguments for as.matrix()}
}
\description{
Returns the embedding matrix. All components can be accessed
using phate$embedding, phate$diff.op, etc
}
