% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{calculate.kernel}
\alias{calculate.kernel}
\title{Calculate a Gaussian or kNN kernel}
\usage{
calculate.kernel(data, k = 15, alpha = 10, npca = 100,
  pca.method = "random", knn.dist.method = "euclidean")
}
\arguments{
\item{data}{The data (n.samples x n.samples)}

\item{k}{number of neighbors}

\item{alpha}{decay constant}

\item{npca}{number of dimensions of PCA reduction to do before KNN}

\item{pca.method}{The desired method for implementing pca for preprocessing the
data. Options include 'svd', 'random', and 'none' (no pca).
Default is 'random'.}
}
\value{
g.kernel Kernel matrix (n.samples x n.samples)
}
\description{
Calculate a Gaussian or kNN kernel
}
