% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{calculate.landmark.operator}
\alias{calculate.landmark.operator}
\title{Calculate the landmark operator}
\usage{
calculate.landmark.operator(g.kernel, n.landmark = 2000, n.svd = 100)
}
\arguments{
\item{g.kernel}{Kernel matrix (n.samples x n.samples)}

\item{n.landmark}{Number of landmarks to compute, or NA. Default is 2000. If NA, then exact PHATE is computed.}

\item{n.svd}{Number of singular vectors to calculate for landmark selection. Default is 100.}

\item{seed}{Integer or NA, integer value for random number generator.}
}
\value{
List containing:
\itemize{
\item \strong{diff.op} The diffusion operator (n.landmark x n.landmark)
\item \strong{landmark.transitions} Transition matrix between cells and landmark (n.cells x n.landmark)
}
}
\description{
Calculate the landmark operator
}
