% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pyphate_version}
\alias{check_pyphate_version}
\title{Check that the current PHATE version in Python is up to date.}
\usage{
check_pyphate_version()
}
\description{
Check that the current PHATE version in Python is up to date.
}
