% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vne.R
\name{find.knee.point}
\alias{find.knee.point}
\title{Returns the x-location of a (single) knee of curve y=f(x)}
\usage{
find.knee.point(y, x = NULL)
}
\arguments{
\item{y}{array, shape=\link{n}. Data for which to find the knee point}

\item{x}{array, optional, shape=\link{n}, default=np.arange(len(y)). Indices of the data points of y, if these are not in order and evenly spaced}
}
\value{
knee_point int The index (or x value) of the knee point on y
}
\description{
Returns the x-location of a (single) knee of curve y=f(x)
}
