% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{ggplot.phate}
\alias{ggplot.phate}
\title{Convert a PHATE object to a data.frame for ggplot}
\usage{
\method{ggplot}{phate}(phate, ...)
}
\arguments{
\item{phate}{A fitted PHATE object}

\item{...}{Arguments for ggplot()}
}
\description{
Passes the embedding matrix to ggplot with column names PHATE1 and PHATE2
}
\examples{
data(tree.data)
phate.tree <- phate(tree.data$data)
ggplot2::ggplot(phate.tree, aes(x=PHATE1, y=PHATE2, color=tree.data$branches)) +
  geom_point()
}
