% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mds}
\alias{mds}
\title{Performs specified method of PCA}
\usage{
mds(data, ndim = 2, method = "metric", dist.method = "euclidean")
}
\arguments{
\item{method}{The desired method for implementing pca for preprocessing the
data. Options include 'svd', 'random', and 'none' (no pca).}

\item{dist.method}{The desired distance function for MDS. Choices are 'euclidean'
and 'cosine'. Default is 'euclidean'.}

\item{X}{Data matrix.}

\item{k}{Number of PCA dimensions returned.}
}
\value{
A matrix containing the embedded data is returned.
}
\description{
Implements different methods of PCA
}
