% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vne.R
\name{optimal.t}
\alias{optimal.t}
\title{Selects the optimal value of t based on the knee point of the
Von Neumann Entropy of the diffusion operator.}
\usage{
optimal.t(diff.op, t.max = 100, plot = FALSE, double.step = FALSE)
}
\arguments{
\item{t.max}{: int, default: 100 Maximum value of t to test}

\item{plot}{: boolean, default: False
If true, plots the Von Neumann Entropy and knee point}

\item{double.step}{: boolean, if true, double the steps in t}
}
\value{
t_opt : int. The optimal value of t
}
\description{
Selects the optimal value of t based on the knee point of the
Von Neumann Entropy of the diffusion operator.
}
