% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{phate}
\alias{phate}
\title{Runs PHATE on an input data matrix}
\usage{
phate(data, t = 20, k = 5, alpha = 10, ndim = 2,
  pca.method = "random", npca = 100, mds.method = "cmds",
  dist.method = "euclidean", mds.dist.method = "euclidean", diff.op = NA,
  diff.op.t = NA)
}
\arguments{
\item{data}{Data matrix. Must have cells on the rows and genes on the columns}

\item{t}{Diffusion time scale. Default is 20.}

\item{k}{k for the adaptive kernel bandwidth. Default is 5.}

\item{alpha}{The alpha parameter in the exponent of the kernel function. Determines the kernel decay rate. Default is 10.}

\item{ndim}{The number of desired PHATE dimensions in the output Y. 2 or 3
is best for visualization. A higher number can be used for
running other analyses on the PHATE dimensions.}

\item{pca.method}{The desired method for implementing pca for preprocessing the
data. Options include 'svd', 'random', and 'none' (no pca).
Default is 'random'.}

\item{npca}{The number of PCA components for preprocessing the data. Default is 100.}

\item{mds.method}{Method for implementing MDS. Choices are 'cmds', 'mmds', and 'nmmds'. Default is 'cmds'.}

\item{dist.method}{The desired distance function for calculating pairwise
distances on the data. Default is 'euclidean'.}

\item{mds.dist.method}{The desired distance function for MDS. Choices are 'euclidean'
and 'cosine'. Default is 'euclidean'.}

\item{diff.op}{If the diffusion operator has been computed on a prior run with the
desired parameters, then this option can be used to directly input the
diffusion operator to save on computational time. Default is NA.}

\item{diff.op.t}{Same as for 'DiffOp', if the powered diffusion operator has been
computed on a prior run with the desired parameters then this
option can be used to directly input the diffusion operator to
save on computational time. Default is NA.}
}
\value{
List containing:
\itemize{
\item \strong{Y}: the PHATE embedding
\item \strong{diff.op}: The diffusion operator which can be used as optional input with another run.
\item \strong{diff.op.t}: diff.op^t
\item \strong{g.kernel}: The kernel used to construct the diffusion operator
}
}
\description{
PHATE is a data reduction method specifically designed for visualizing \strong{high}
dimensional data in \strong{low} dimensional spaces.
}
