% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{plot.phate}
\alias{plot.phate}
\title{Plot a PHATE object in base R}
\usage{
\method{plot}{phate}(phate, ...)
}
\arguments{
\item{phate}{A fitted PHATE object}

\item{...}{Arguments for plot()}
}
\description{
Plot a PHATE object in base R
}
\examples{
data(tree.data)
phate.tree <- phate(tree.data$data)
plot(phate.tree, col=tree.data$branches))
}
