% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{plot.phate}
\alias{plot.phate}
\title{Plot a PHATE object in base R}
\usage{
\method{plot}{phate}(x, ...)
}
\arguments{
\item{x}{A fitted PHATE object}

\item{...}{Arguments for plot()}
}
\description{
Plot a PHATE object in base R
}
\examples{
if (reticulate::py_module_available("phate")) {

library(graphics)
# data(tree.data)
# We use a smaller tree to make examples run faster
data(tree.data.small)
phate.tree <- phate(tree.data.small$data)
plot(phate.tree, col=tree.data.small$branches)

}
}
