% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{print.phate}
\alias{print.phate}
\title{Print a PHATE object}
\usage{
\method{print}{phate}(phate, ...)
}
\arguments{
\item{phate}{A fitted PHATE object}

\item{...}{Arguments for print()}
}
\description{
This avoids spamming the user's console with a list of many large matrices
}
\examples{
data(tree.data)
phate.tree <- phate(tree.data$data)
print(phate.tree)
## PHATE embedding with elements
##   $embedding : (3000, 2)
##   $g.kernel : (3000, 3000)
##   $diff.op : (2000, 2000)
##   $diff.op.t : (2000, 2000)
##   $params : list with elements (data, k, alpha, t, n.landmark, 
##                                 potential.method, pca.method, 
##                                 npca, n.svd, mds.method, 
##                                 knn.dist.method, mds.dist.method)
}
