% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{summary.phate}
\alias{summary.phate}
\title{Summarize a PHATE object}
\usage{
\method{summary}{phate}(object, ...)
}
\arguments{
\item{object}{A fitted PHATE object}

\item{...}{Arguments for summary()}
}
\description{
Summarize a PHATE object
}
\examples{
if (reticulate::py_module_available("phate")) {

data(tree.data)
phate.tree <- phate(tree.data$data)
summary(phate.tree)
## PHATE embedding
## k = 5, alpha = NA, t = 58
## Data: (3000, 100)
## Embedding: (3000, 2)

}
}
