% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdpca.R
\name{svdpca}
\alias{svdpca}
\title{Performs specified method of PCA}
\usage{
svdpca(X, k, method)
}
\arguments{
\item{X}{Data matrix.}

\item{k}{Number of PCA dimensions returned.}

\item{method}{The desired method for implementing pca for preprocessing the
data. Options include 'svd', 'random', and 'none' (no pca).}
}
\value{
A matrix containing the rotated data is returned.
}
\description{
Implements different methods of PCA
}
