% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.data.R
\docType{data}
\name{tree.data}
\alias{tree.data}
\title{Fake branching data for examples}
\format{A list containing \code{data}, a matrix with 3000 rows and 100 variables
and \code{branches}, a factor containing 3000 elements.}
\source{
The authors
}
\usage{
tree.data
}
\description{
A dataset containing high dimensional data that has 10 unique branches
}
\keyword{datasets}
