% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.data.R
\docType{data}
\name{tree.data.small}
\alias{tree.data.small}
\title{Fake branching data for running examples fast}
\format{
A list containing \code{data}, a matrix with 250 rows and 50 variables
and \code{branches}, a factor containing 250 elements.
}
\source{
The authors
}
\usage{
tree.data.small
}
\description{
A dataset containing high dimensional data that has 10 unique branches
}
\keyword{datasets}
