% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vne.R
\name{vne}
\alias{vne}
\title{Determines the Von Neumann entropy of data
at varying matrix powers. The user should select a value of t
around the "knee" of the entropy curve.}
\usage{
vne(diff.op, t.max = 200)
}
\arguments{
\item{t.max}{: int, default=200. Maximum value of t to test}
}
\value{
entropy : array, shape=(t_max).
The entropy of the diffusion affinities for each value of t
}
\description{
Determines the Von Neumann entropy of data
at varying matrix powers. The user should select a value of t
around the "knee" of the entropy curve.
}
