% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phate.R
\name{phate}
\alias{phate}
\title{Run PHATE on an input data matrix}
\usage{
phate(data, ndim = 2, t = "auto", k = 5, alpha = 10, use.alpha = NA,
  n.landmark = 2000, potential.method = "log", t.max = 100,
  pca.method = "random", npca = 100, n.svd = 100, mds.method = "metric",
  knn.dist.method = "euclidean", mds.dist.method = "euclidean",
  init = NULL, verbose = TRUE, plot.optimal.t = FALSE, g.kernel = NULL,
  diff.op = NULL, landmark.transitions = NULL, diff.op.t = NULL,
  dist.method = NA)
}
\arguments{
\item{data}{matrix (n_samples, n_dimensions)
2 dimensional input data array with
n_samples samples and n_dimensions dimensions}

\item{ndim}{int, optional, default: 2
number of dimensions in which the data will be embedded}

\item{t}{int, optional, default: 'auto'
power to which the diffusion operator is powered
sets the level of diffusion}

\item{k}{int, optional, default: 5
number of nearest neighbors on which to build kernel}

\item{alpha}{int, optional, default: NA
sets decay rate of kernel tails.
If NA, alpha decaying kernel is not used}

\item{use.alpha}{boolean, default: NA
forces the use of alpha decaying kernel
If NA, alpha decaying kernel is used for small inputs
(n_samples < n_landmark) and not used otherwise}

\item{n.landmark}{int, optional, default: 2000
number of landmarks to use in fast PHATE}

\item{potential.method}{string, optional, default: 'log'
choose from 'log' and 'sqrt'
which transformation of the diffusional operator is used
to compute the diffusion potential}

\item{t.max}{int, optional, default: 100.
Maximum value of t to test for automatic t selection.}

\item{pca.method}{string, optional, default: 'random'
The desired method for implementing pca for preprocessing the
data. Options include 'svd', 'random', and 'none' (no pca).}

\item{npca}{int, optional, default: 100
Number of principal components to use for calculating
neighborhoods. For extremely large datasets, using
n_pca < 20 allows neighborhoods to be calculated in
log(n_samples) time.}

\item{n.svd}{int, optional, default: 100
The number of SVD components to use for landmark selection}

\item{mds.method}{string, optional, default: 'metric'
choose from 'classic', 'metric', and 'nonmetric'
which MDS algorithm is used for dimensionality reduction}

\item{knn.dist.method}{string, optional, default: 'euclidean'.
The desired distance function for calculating pairwise distances on the data.}

\item{mds.dist.method}{string, optional, default: 'euclidean'
recommended values: 'euclidean' and 'cosine'}

\item{init}{phate object, optional
object to use for initialization. Avoids recomputing
intermediate steps if parameters are the same.}

\item{verbose}{boolean, optional, default : TRUE
If TRUE, print verbose updates.}

\item{plot.optimal.t}{boolean, optional, default: FALSE
If TRUE, produce a plot showing the Von Neumann Entropy
curve for automatic t selection.}

\item{g.kernel}{Precomputed kernel matrix}

\item{diff.op}{Precomputed diffusion operator}

\item{landmark.transitions}{Precomputed landmark transitions}

\item{diff.op.t}{Precomputed powered diffusion operator}
}
\value{
"phate" object containing:
\itemize{
\item \strong{embedding}: the PHATE embedding
\item \strong{diff.op}: The diffusion operator which can be used as optional input with another run.
\item \strong{diff.op.t}: diff.op^t
\item \strong{g.kernel}: The kernel used to construct the diffusion operator
\item \strong{params}: Parameters passed to phate
}
}
\description{
PHATE is a data reduction method specifically designed for visualizing \strong{high}
dimensional data in \strong{low} dimensional spaces.
}
