% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.data.R
\docType{data}
\name{tree.data}
\alias{tree.data}
\title{Fake branching data for examples}
\format{A data.frame with 1440 rows and 61 variables}
\source{
The authors
}
\usage{
tree.data
}
\description{
A dataset containing high dimensional data that has branchpoints
}
\keyword{datasets}
