% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vne.R
\name{vne}
\alias{vne}
\title{Determines the Von Neumann entropy of data
at varying matrix powers. The user should select a value of t
around the "knee" of the entropy curve.}
\usage{
vne(diff.op, t.max = 100)
}
\arguments{
\item{t.max}{: int, default=100. Maximum value of t to test}
}
\value{
entropy : array, shape=(t_max).
The entropy of the diffusion affinities for each value of t
}
\description{
Determines the Von Neumann entropy of data
at varying matrix powers. The user should select a value of t
around the "knee" of the entropy curve.
}
\examples{
X=diag(nrow=10)
X[1,1] <- 5
X[3,4] <- 4
stopifnot(find.knee.point(vne(X, t.max=100)) == 24)
}
